requirements.txt
streamlit
google-generativeai
altair